Pull all paginated resources from prismic API
A patched metalsmith-prismic fork was added as a local node dependency.
This pulls all pages of paginated API resources from prismic. For now
it was added to the static site generator's git repo. Potentially this
could be pushed upstream to the original repo via a GitHub pull request
on https://github.com/mbanting/metalsmith-prismic