Limit number of blog posts on front page to three
The `allPages` parameter was introduced to metalsmith-prismic.
Setting it to true means all pages from a paginated resources get
pulled. To make this a non-breaking change, it defaults to
false/undefined. For our use case it should rather default to true, but
this way the change can be push upstream to the original library on
github while keeping it backwards compatible.